+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Master
 
+### Minor
+
+- Give access to the base of each module.
+
 ### Patch
 
 - Fix missing panics in documentation.